requirements.txt
streamlit==1.32.0
openai==1.12.0
chromadb==0.4.22
python-dotenv==1.0.1
langchain==0.1.9
streamlit-calendar==1.1.0
SpeechRecognition==3.10.0
gTTS==2.5.0
python-dateutil==2.8.2
pyttsx3==2.90
PyAudio==0.2.13
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.86.0